Switch from unsigned int to long int

+++ gene_bin.h
@@ -33,27 +33,27 @@ typedef struct mutation_map {
 
 /********** BINARIES FUNCTION **********/
 
-int get_binary_value(const unsigned int *seq_bin, const int pos);
-unsigned int* change_binary_value(unsigned int *seq_bin, const int pos, const int value);
-unsigned int* set_binary_array(const char *array, const unsigned size);
-unsigned int* xor_binary_array(const unsigned int *seq1, const unsigned array_size1,
-                                const unsigned int *seq2, const unsigned array_size2);
-int popcount_binary_array(const unsigned int *seq, const unsigned int size);
-unsigned int mask_binary_array(const unsigned int seq_bin, const unsigned pos_start, const unsigned size);
-unsigned int* get_piece_binary_array(const unsigned int* seq_bin, const unsigned int seq_size, const unsigned pos_start, const unsigned size);
+int get_binary_value(const long int *seq_bin, const int pos);
+long int* change_binary_value(long int *seq_bin, const int pos, const int value);
+long int* set_binary_array(const char *array, const unsigned size);
+long int* xor_binary_array(const long int *seq1, const unsigned array_size1,
+                                const long int *seq2, const unsigned array_size2);
+int popcount_binary_array(const long int *seq, const long int size);
+long int mask_binary_array(const long int seq_bin, const long int pos_start, const long int size);
+long int* get_piece_binary_array(const long int* seq_bin, const long int seq_size, const long int pos_start, const long int size);
 
 
 /******** DNA & GENES FUNCTION *********/
 
-unsigned int* convert_to_binary(const char* dna_seq, const unsigned size);
-char* binary_to_dna(unsigned int* bin_dna_seq, const unsigned size);
-char* generating_mRNA(const unsigned int* gene_seq, const unsigned int seq_size);
-void detecting_genes(const unsigned int *gene, const unsigned int gene_size,
+long int* convert_to_binary(const char* dna_seq, const unsigned size);
+char* binary_to_dna(long int* bin_dna_seq, const unsigned size);
+char* generating_mRNA(const long int* gene_seq, const long int seq_size);
+void detecting_genes(const long int *gene, const long int gene_size,
                      gene_map_t* gene_map);
-char* generating_amino_acid_chain(const unsigned int *gene_seq, const unsigned int seq_size);
-void detecting_mutations(const unsigned int *gene_seq, const unsigned int size_sequence,
+char* generating_amino_acid_chain(const long int *gene_seq, const long int seq_size);
+void detecting_mutations(const long int *gene_seq, const long int size_sequence,
                          mutation_map mut_m);
-float calculating_matching_score(const unsigned int *seq1, const int sequence_size1,
-                                 const unsigned int *seq2, const int sequence_size2);
+float calculating_matching_score(const long int *seq1, const int sequence_size1,
+                                 const long int *seq2, const int sequence_size2);
